pipeline.plot: refactor code into separate files where applicable
