+++ pipeline/plot/__init__.py
@@ -1,99 +1,6 @@
-#!/usr/bin/env python
+# -*- mode: python -*-
 
-import os
-import inspect 
+from .util import show_source
 
-import numpy as np
-import scipy as sp
-import datajoint as dj
-
-import matplotlib.pyplot as plt
-
-from IPython.core.display import HTML, display
-from pygments import highlight
-from pygments.lexers import PythonLexer
-from pygments.formatters import HtmlFormatter
-
-from pipeline import experiment
-from pipeline import ephys
-from pipeline import psth
-
-def show_source(function):
-    code = inspect.getsource(function)
-    style = HtmlFormatter().get_style_defs('.highlight')
-    html = highlight(code, PythonLexer(), HtmlFormatter(style='colorful'))
-    display(HTML('<style>{}</style>'.format(style)))
-    display(HTML(html))
-
-
-def plot_unit_data(ipsi_hit, contra_hit, ipsi_err, contra_err):
-    max_trial_off = 500
-    binSize=0.04
-    plt_xmin=-3
-    plt_xmax=3
-    plt_ymin=0
-    plt_ymax=None  # dynamic per unit
-    
-    plt_ymax = np.max([contra_hit['psth'][0], ipsi_hit['psth'][0], contra_err['psth'][0], ipsi_err['psth'][0]])
-    
-    plt.figure()
-    
-    # raster plot
-    ax=plt.subplot(411)
-    plt.plot(contra_hit['raster'][0], contra_hit['raster'][1] + max_trial_off, 'b.', markersize=1)
-    plt.plot(ipsi_hit['raster'][0], ipsi_hit['raster'][1], 'r.', markersize=1)
-    ax.set_axis_off()
-    ax.set_xlim([plt_xmin, plt_xmax])
-    ax.axvline(0,0,1, ls='--')
-    ax.axvline(-1.2,0,1, ls='--')
-    ax.axvline(-2.4,0,1, ls='--')
-
-    # histogram of hits
-    ax = plt.subplot(412)
-    plt.plot(contra_hit['psth'][1][1:], contra_hit['psth'][0], 'b')
-    plt.plot(ipsi_hit['psth'][1][1:], ipsi_hit['psth'][0], 'r')
-    
-    plt.ylabel('spikes/s')
-    ax.spines["top"].set_visible(False)  
-    ax.spines["right"].set_visible(False)
-    ax.set_xlim([plt_xmin, plt_xmax])
-    ax.set_ylim([plt_ymin, plt_ymax])
-    ax.set_xticklabels([])
-    ax.axvline(0,0,1, ls='--')
-    ax.axvline(-1.2,0,1, ls='--')
-    ax.axvline(-2.4,0,1, ls='--')
-    plt.title('Correct trials')
- 
-    # histogram of errors
-    ax = plt.subplot(413)
-    plt.plot(contra_err['psth'][1][1:], contra_err['psth'][0], 'b')
-    plt.plot(ipsi_err['psth'][1][1:], ipsi_err['psth'][0], 'r')
-    
-    ax.spines["top"].set_visible(False)  
-    ax.spines["right"].set_visible(False)
-    ax.set_xlim([plt_xmin, plt_xmax])
-    ax.set_ylim([plt_ymin, plt_ymax])
-    ax.axvline(0,0,1, ls='--')
-    ax.axvline(-1.2,0,1, ls='--')
-    ax.axvline(-2.4,0,1, ls='--')
-
-    plt.title('Error trials')
-    plt.xlabel('Time to go cue (s)')
-    plt.show()
-
-
-def plot_unit(unit_key):
-    
-    ipsi_hit_cond_key = (psth.Condition() & {'condition_desc': 'audio delay ipsi hit'}).fetch1('KEY')
-    contra_hit_cond_key = (psth.Condition() & {'condition_desc': 'audio delay contra hit'}).fetch1('KEY')
-
-    ipsi_err_cond_key = (psth.Condition() & {'condition_desc': 'audio delay ipsi error'}).fetch1('KEY')
-    contra_err_cond_key = (psth.Condition() & {'condition_desc': 'audio delay contra error'}).fetch1('KEY')
-    
-    ipsi_hit_cell_psth = psth.CellPsth.get(ipsi_hit_cond_key, unit_key)
-    contra_hit_cell_psth = psth.CellPsth.get(contra_hit_cond_key, unit_key)
-
-    ipsi_err_cell_psth = psth.CellPsth.get(ipsi_hit_cond_key, unit_key)
-    contra_err_cell_psth = psth.CellPsth.get(contra_err_cond_key, unit_key)
-    
-    plot_unit_data(ipsi_hit_cell_psth, contra_hit_cell_psth, ipsi_err_cell_psth, contra_err_cell_psth)
+from .cell_psth import unit_psth
+from .cell_psth import unit_psth_ll
